Fix all examples to run either from root or in examples dir
Fixed all errors in examples that were due to out dated code (pandas
issues with options and indexing). Also fixed it so that all the paths
that use git-pandas as the repo do it in a way that they can be ran
inside the examples directory or at the root directory. Part of doing
this was to better setup the setup.py file. All the requirements are now
in the setup.py, and the requirements.txt file says to `-e .` so that it
install git-pandas in the local envirornment in editable mode. The
normal reccomended way for package development. And then also added an
optional extra dependencies section and added examples dependencies like
matplotlib, so you can install those by going pip isntall
git-pandas[examples]. Also added a requirements-dev.txt which installs
the requirements.txt, but also the optional extra dependencies, so you
can do things like run the examples.